refactor checkbox and radio button to reuse CheckInput
