+++ src/modules/coreUI/components/basic/RadioButton.js
@@ -1,66 +1,30 @@
 /* eslint-disable indent */
 import React from 'react';
-import styled from 'react-emotion';
-import { css } from 'emotion';
 import PropTypes from 'prop-types';
-import withMedia from '~/modules/core/utils/mediaHelpers/withMedia';
-import { Row } from '~/modules/coreUI/components/layouts/helpers/LinearLayout';
-import Spacer from '~/modules/coreUI/components/layouts/helpers/Spacer';
-import { propsForPrefix, infereControlType, infereNamedFontSize, responsiveStyle, infereBorderRadius, colorStyles, disabledColorStyles } from '~/modules/coreUI/utils/infereStyle';
 import Icon from '~/modules/coreUI/components/basic/Icon';
-import spaceStyles from '~/modules/coreUI/utils/styleSystem';
+import CheckInput from './CheckInput';
 
-import { Label } from './Labels';
-
-const RootContainer = styled(Row)`
-  cursor: pointer;
-`;
-
-const RealHiddenRadio = styled.input`
-  opacity: 0;
-  position: absolute;
-  left: -999px;
-
-  & + div {
-    ${props => responsiveStyle(props, 'size', size => css`
-      /* Workaround for Flexbox & Grid inconsistencies when using only width/height */
-      min-width: ${1.3 * infereNamedFontSize(props, size)}px;
-      max-width: ${1.3 * infereNamedFontSize(props, size)}px;
-      min-height: ${1.3 * infereNamedFontSize(props, size)}px;
-      max-height: ${1.3 * infereNamedFontSize(props, size)}px;
-      font-size: ${0.7 * infereNamedFontSize(props, size)}px;
-    `)};
-
-    border: 1px solid;
-    border-radius: ${props => infereBorderRadius(props)}px; 
-    
-    ${props => spaceStyles(props)}
-    ${props => (props.disabled ? disabledColorStyles(props) : colorStyles(props))}
-  }
-
-  &:focus + div {
-    border-color: ${props => props.theme.new.inputs.focusBorderColor[infereControlType(props)]};
-  }
-`;
-
-// helper function to copy static method from component to another comopnent
-// withMedia HOC used bellow don't copy static method
-// this function do so
-// a better solution will be to use this https://github.com/mridgway/hoist-non-react-statics
-// but for this simple case i will use this function
-function setStaticMethodToComponent(targetComponent, sourceComponent) {
-  /* eslint-disable no-param-reassign */
-  // note this function is not pure function
-  // it contain a side effect targetComponent reference will change its property
-  return Object
-    .entries(sourceComponent)
-    .reduce((newTargetComponent, [key, value]) => {
-      if (!newTargetComponent[key]) {
-        newTargetComponent[key] = value;
-      }
-      return newTargetComponent;
-    }, targetComponent);
-}
+/**
+ * we don't need this function now 😊
+ */
+// // helper function to copy static method from component to another comopnent
+// // withMedia HOC used bellow don't copy static method
+// // this function do so
+// // a better solution will be to use this https://github.com/mridgway/hoist-non-react-statics
+// // but for this simple case i will use this function
+// function setStaticMethodToComponent(targetComponent, sourceComponent) {
+//   /* eslint-disable no-param-reassign */
+//   // note this function is not pure function
+//   // it contain a side effect targetComponent reference will change its property
+//   return Object
+//     .entries(sourceComponent)
+//     .reduce((newTargetComponent, [key, value]) => {
+//       if (!newTargetComponent[key]) {
+//         newTargetComponent[key] = value;
+//       }
+//       return newTargetComponent;
+//     }, targetComponent);
+// }
 
 const isFunction = f => typeof f === 'function';
 
@@ -74,44 +38,14 @@ function RadioButton(props) {
       {({ value: selectedValue, setValue }) => {
         const checked = value === selectedValue;
         return (
-          <RootContainer
-            centerAligned
+          <CheckInput
+            {...props}
+            type="radio"
+            checked={checked}
             onClick={() => !disabled && setValue(value)}
-          >
-            {/* i think onChange will never be called but i see it on orignal example */}
-            <RealHiddenRadio
-              type="radio"
-              id={props.elemID}
-              {...props}
-              inverted={props.inverted || !checked}
-              checked={checked}
-              onChange={() => setValue(value)}
-            />
-            <Row
-              fullWidth
-              fullHeight
-              centerAligned
-              centerJustified
-              type="radio"
-              {...props}
-            >
-              {checked &&
-                <Icon inheritColor inheritSize className="fas fa-circle" />
-              }
-            </Row>
-            {props.label &&
-              <React.Fragment>
-                <Spacer />
-                <Label
-                  size={props.size}
-                  bold={props.bold}
-                  {...propsForPrefix(props, 'label_')}
-                >
-                  {props.label}
-                </Label>
-              </React.Fragment>
-            }
-          </RootContainer>
+            onChange={() => setValue(value)}
+            renderIcon={() => <Icon className="fas fa-circle" />}
+          />
         );
       }}
     </RadioButtonContext.Consumer>
@@ -130,10 +64,8 @@ RadioButton.Group = class Group extends React.Component {
   }
   // if this component is controlled then we need to use value from props
   // insted of our Component state
-  getState() {
-    return {
-      value: (this.isControlled() ? this.props : this.state).value,
-    };
+  getValue() {
+    return (this.isControlled() ? this.props : this.state).value;
   }
   setValue = (value) => {
     if (this.isControlled()) {
@@ -147,7 +79,7 @@ RadioButton.Group = class Group extends React.Component {
   }
   getHelpers() {
     return {
-      ...this.getState(),
+      value: this.getValue(),
       setValue: this.setValue,
     };
   }
@@ -175,4 +107,4 @@ RadioButton.Group.propTypes = PropTypes.shape({
   defaultValue: PropTypes.any,
 }).isRequired;
 
-export default setStaticMethodToComponent(withMedia(RadioButton), RadioButton);
+export default RadioButton;
